switch PI control to float arithmetic

+++ inc/main.h
@@ -154,13 +154,13 @@ typedef struct
 
 typedef struct
 {
-	int16_t       	gain_p;
-	int16_t       	gain_i;
+	float       	gain_p;
+	float       	gain_i;
 	int16_t       	limit_i;
 	int16_t       	limit_output;
 	int16_t       	recent_value;
 	int32_t       	setpoint;
-	int32_t       	integral_part;
+	float       	integral_part;
 	int16_t       	max_step;
 	int32_t       	out;
 	int8_t       	shift;
